==> 03-RNAseq/03c-stringtie_curate.ipynb <==
# Filtering stringtie predictions

Because fungal genes are so close together, it's fairly common for RNA-seq coverage to overlap between neighboring genes.
As a consequence, stringtie tends to create alternative transcripts that are actually just merges of two genes.
This is the problem we're trying to resolve here.

import re
import json
import pandas as pd
import numpy as np
from scipy.sparse import lil_array
from collections import defaultdict, Counter
from intervaltree import Interval, IntervalTree

from scipy.spatial.distance import squareform
from scipy.cluster.hierarchy import linkage, dendrogram, cut_tree

from IPython.display import clear_output

LINEWIDTH=160

pd.set_option('display.width', LINEWIDTH)

regex1 = re.compile('transcript_id\s+"([^"]+)"')
regex2 = re.compile('ref_gene_id\s+"(?P<g>[^"]+)"')
regex3 = re.compile('gene_id\s+"([^"]+)"')

gtf = pd.read_csv(
    "stringtie_merged.gtf",
    sep="\t",
    comment="#",
    names="seqid source type start end score strand phase attributes".split(),
    low_memory=False
)

def get_ref_gene_id(f):
    m = regex2.search(f)

    if m is None:
        return pd.NA

    return m.groupdict().get("g")

gtf["start"] -= 1
gtf["transcript"] = gtf["attributes"].apply(lambda f: regex1.search(f).groups()[0])
gtf["gene_id"] = gtf["attributes"].apply(lambda f: regex3.search(f).groups()[0])
gtf["ref_gene_id"] = gtf["attributes"].apply(get_ref_gene_id)
gtf.drop(columns = "attributes", inplace=True)
gtf.sort_values(by=["seqid", "start", "end"], inplace=True)
gtf.head()

First, we'll need a way to find overlapping genes.
This is relatively simple if we go by the gene names, but later on when we split genes this will be more useful.

gtf["transcript"].nunique()

itrees = defaultdict(IntervalTree)
transcript_indices = dict()

for i, row in gtf[gtf["type"] == "transcript"].reset_index().iterrows():
    transcript = row["transcript"]
    transcript_indices[transcript] = i
    itrees[row["seqid"]].add(Interval(row["start"], row["end"], data=transcript))

Next we need a way of comparing two gene predictions.
We'll define the similarity as the jaccard index of the exon positions (ie. how similar is the overall intron-exon structure).
We'll also look as locus level merging based on a simple overlap coverage.

Ideally, we shouldn't have any cases where two loci that don't overlap, are covered by another locus that overlaps both.

transcripts = defaultdict(set)

for i, row in gtf[gtf["type"] == "transcript"].iterrows():
    transcript = row["transcript"]
    if row["strand"] == "-":
        strand = -1
        offset = -1
    else:
        strand = 1
        offset = 0

    transcripts[transcript].update(range((strand * row["start"]) + offset, (strand * row["end"]) + offset, strand))


exons = defaultdict(set)

for i, row in gtf[gtf["type"] == "exon"].iterrows():
    transcript = row["transcript"]
    if row["strand"] == "-":
        strand = -1
    else:
        strand = 1

    exons[transcript].update(range(strand * row["start"], strand * row["end"], strand))

def coverage(t1, t2):
    inter = len(t1.intersection(t2))
    return max([inter / len(t1), inter / len(t2)])

Now we'll calculate the pairwise coverages.

coverages = lil_array((len(transcript_indices), len(transcript_indices)), dtype=float)

for seqid, itree in itrees.items():
    for interval in sorted(itree):
        t1_name = interval.data
        t1_index = transcript_indices[t1_name]
        t1 = transcripts[t1_name]
        e1 = exons[t1_name]
        overlapping = itree.overlap(interval)
        for overlap in sorted(overlapping):
            t2_name = overlap.data
            t2_index = transcript_indices[t2_name]
            t2 = transcripts[t2_name]
            cov = coverage(t1, t2)
            coverages[t1_index, t2_index] = cov

From this we get a way to quickly compare two genes.

coverages[:5, :5].todense()

## Manual removal of loci that join distinct loci

Here we'll do a semi-manual curation of the transcripts.
Some cases are easy, while others are more in a grey area.
We'll look at the cases where it's a grey area and automatically deal with the easy ones.

- Joining transcripts are the problematic transcripts that we want to exclude.
- Joining exceptions are the transcripts that may look like problems, but we actually want to keep them in.

def is_bad(subdf, transcript_indices, coverages, good_prefix = "sscl"):
    gene_transcripts = subdf["transcript"].unique()
    gene_transcripts_indices = np.array([transcript_indices[t] for t in gene_transcripts])
    
    goodies = np.array([s.lower().startswith(good_prefix) for s in gene_transcripts])
    
    n = gene_transcripts.shape[0]
    col_index = np.repeat(gene_transcripts_indices, n)
    row_index = np.tile(gene_transcripts_indices, n)

    transcript_cov = np.reshape(
        coverages[row_index, col_index].todense(),
        (n, n)
    )
    
    transcript_cov = pd.DataFrame(transcript_cov, index=gene_transcripts, columns=gene_transcripts)

    if transcript_cov.size == 1:
        return transcript_cov, gene_transcripts, None, None

    if (transcript_cov.values < 1).any(axis=1).any():
        easy_cases = gene_transcripts[(transcript_cov.values > 0.99).all(axis=1) & (~goodies)]
        baddies = gene_transcripts[(transcript_cov.values > 0).all(axis=1) & (~goodies)]

        if np.size(easy_cases) == 0:
            easy_cases = None
        else:
            baddies = np.setdiff1d(baddies, easy_cases)
        
        if np.size(baddies) == 0:
            baddies = None
    else:
        baddies = None
        easy_cases = None

    
    return transcript_cov, gene_transcripts, baddies, easy_cases

def plot_exons(subdf, mapper=None, step=10, linewidth=80):
    subdf_exons = subdf[subdf["type"] == "exon"]
    start = subdf_exons["start"].min()
    end = subdf_exons["end"].max()

    seqs = dict()
    seqs["loc"] = "".join(f"{d: <10}" for d in range(start, end - step + 1, step))
    seqs["markers"] = "".join(["|" if i % 10 == 0 else " " for i in range(len(seqs["loc"]))])

    for tid, subdf2 in subdf.groupby("transcript"):
        subdf2_exons = subdf2[subdf2["type"] == "exon"]
        seq = np.array([" " for _ in range(end - start)])

        last_end = None
        for i, e in subdf2_exons.iterrows():
            seq[e["start"] - start: e["end"] - start] = e["strand"][0]
            if last_end is None:
                last_end = e["end"] - start
            else:
                seq[last_end: e["start"] - start] = "i"

            

        seq = "".join(seq)

        compressed_string = []
        for i in range(0, len(seq) - step + 1, step):
            p = seq[i:i + step].count("+")
            m = seq[i:i + step].count("-")
            g = seq[i:i + step].count(".")
            i = seq[i:i + step].count("i")

            if g >= p and g >= m and g >= i:
                compressed_string.append(".")
            elif i >= p and i >= m and i >= g:
                compressed_string.append("=")
            elif p >= g and p >= m and p >= i:
                compressed_string.append(">")
            elif m >= g and m >= p and m >= i:
                compressed_string.append("<")
            else:
                compressed_string.append("?")

        seqs[tid] = "".join(compressed_string)

    if mapper is not None:
        max_mapper_len = max([len(de) for de in mapper.values()]) + 2
    else:
        max_mapper_len = 0
        mapper = {id_: "" for id_ in seqs.keys()}
    
    max_id_len = max([len(id_) for id_ in seqs.keys()])
    
    for start in range(0, len(compressed_string) -  1, linewidth):
        end = start + linewidth
    
        for id_, seq in seqs.items():
            decision = mapper.get(id_, "")
            print(f"{decision:<{max_mapper_len}}{id_: >{max_id_len}}  ", seq[start:end])

        print()

def plot_transcripts(subdf, mapper=None, step=100, linewidth=80):
    subdf_transcripts = subdf[subdf["type"] == "transcript"]
    start = subdf_transcripts["start"].min()
    end = subdf_transcripts["end"].max()

    seqs = dict()
    seqs["loc"] = "".join(f"{d: <10}" for d in range(start, end - step + 1, step))
    seqs["markers"] = "".join(["|" if i % 10 == 0 else " " for i in range(len(seqs["loc"]))])

    for tid, subdf2 in subdf.groupby("transcript"):

        subdf2_transcript = subdf2[subdf2["type"] == "transcript"]
        seq = np.array([" " for _ in range(end - start)])

        for i, e in subdf2_transcript.iterrows():
            seq[e["start"] - start: e["end"] - start] = e["strand"][0]
            
        seq = "".join(seq)

        compressed_string = []
        for i in range(0, len(seq) - step + 1, step):
            p = seq[i:i + step].count("+")
            m = seq[i:i + step].count("-")
            g = seq[i:i + step].count(".")

            if g >= p and g >= m:
                compressed_string.append(".")
            elif p > g and p > m:
                compressed_string.append(">")
            elif m > g and m > p:
                compressed_string.append("<")
            else:
                raise ValueError("this shouldn't happen")

        seqs[tid] = "".join(compressed_string)

    if mapper is not None:
        max_mapper_len = max([len(de) for de in mapper.values()]) + 2
    else:
        max_mapper_len = 0
        mapper = {id_: "" for id_ in seqs.keys()}
    
    max_id_len = max([len(id_) for id_ in seqs.keys()])
    
    for start in range(0, len(compressed_string) -  1, linewidth):
        end = start + linewidth
    
        for id_, seq in seqs.items():
            decision = mapper.get(id_, "")
            print(f"{decision:<{max_mapper_len}}{id_: >{max_id_len}}  ", seq[start:end])

def get_bad_order(transcript_cov, bad_indices, good_prefix="sscl"):
    id_order = transcript_cov.sum(axis=1).sort_values(ascending=True).index
    id_order = id_order[~pd.isna(id_order)].tolist()
    if bad_indices is None:
        id_order = [
            c
            for c
            in transcript_cov
            if not c.lower().startswith(good_prefix)
        ]
    else:
        id_order = [
            c
            for c
            in transcript_cov
            if ((c in bad_indices) and not c.lower().startswith(good_prefix))
        ]

    return id_order

def get_response(worst, subdf=None, linewidth=80):
    response = None
    
    worst = list(reversed(worst))
    while response not in ("y", "n"):
        if len(worst) == 1:
            print(f"\nDelete [y] or keep [n] the worst option {worst[0]}.")
            print("You can also print the exon structure with [p].")
        else:
            print(f"\nDelete [y] or keep [n] the worst option {worst[0]}, select an index, keep all [r], or delete all [a].")
            print("You can also print the exon structure with [p].")
            for h, i in enumerate(worst):
                print(f"- {h}: {i}")

        response = input("\nresponse: ")
        response = response.strip().lower()

        try:
            response_int = int(response)
        except:
            response_int = None
        
        if response in ["n", "k"]:
            return worst[0], "keep"
        elif response in ["y", "d"]:
            return worst[0], "delete"
        elif response in ["a", "da"]:
            return worst, "delete_all"
        elif response in ["r", "ka"]:
            return worst, "keep_all"
        elif response in ["e", "p", "print", "exons"]:
            assert subdf is not None, "If printing exons you need to have provided the subdf"
            print("\nExons:")
            plot_exons(subdf, step=10, linewidth=linewidth)
            return get_response(worst, subdf=subdf, linewidth=linewidth)

        elif (response in worst) or ((response_int is not None) and (response_int < len(worst))):
            if response_int is not None:
                response = worst[response_int]

            response2 = "h"
            while response2 not in ("y", "n", "k", "d", "r"):
                response2 = input(f"\ndelete {response} [y], keep [n], or go back and reselect [r]?: ")

            if response2 in ("n", "k"):
                return response, "keep"
            elif response2 in ("y", "d"):
                return response, "delete"

joining_exceptions = set()
joining_transcripts = set()

for gene_id, subdf in gtf.groupby("gene_id"):
    subdf = subdf[~subdf["transcript"].isin(joining_transcripts | joining_exceptions)]
    transcript_cov, gene_transcripts, baddies, easy_cases = is_bad(subdf, transcript_indices, coverages, good_prefix = "sscl")

    if easy_cases is not None:
        joining_transcripts.update(easy_cases)

    if baddies is None:
        continue
    elif np.size(baddies) == 0:
        continue
    
    baddies = np.setdiff1d(baddies, joining_exceptions)
    if baddies is None:
        continue

    bad_order = get_bad_order(transcript_cov, baddies, good_prefix="sscl")

    # Exclude the easy cases
    subdf = subdf[~subdf["transcript"].isin(joining_transcripts | joining_exceptions)]
    
    while len(bad_order) > 0 and (baddies is not None):
        
        current_baddie = bad_order[-1]

        print(f"# Processing: {gene_id}")

        #print(
        #    subdf["transcript type start end".split()]
        #        .sort_values("transcript start end".split())
        #        .set_index("transcript")
        #)
        
        if easy_cases is not None:
            print("\nAlready excluded easy cases:")
            for e in easy_cases:
                print("-", e)

        print("\nCoverage:")
        print(transcript_cov)
        print()
        plot_transcripts(subdf, step=100, linewidth=LINEWIDTH)
            
        response, action = get_response(baddies, subdf)

        if action == "delete":
            joining_transcripts.add(response)
        elif action == "delete_all":
            joining_transcripts.update(response)
        elif action == "keep":
            joining_exceptions.add(response)
        elif action == "keep_all":
            joining_exceptions.update(response)
        else:
            print(response, action)
            raise ValueError("failed")

        subdf = subdf[~subdf["transcript"].isin(joining_transcripts | joining_exceptions)]
        transcript_cov, gene_transcripts, baddies, easy_cases = is_bad(
            subdf,
            transcript_indices,
            coverages,
            good_prefix = "sscl"
        )

        if easy_cases is not None:
            joining_transcripts.update(easy_cases)

        if baddies is None:
            continue
        elif np.size(baddies) == 0:
            continue

        baddies = np.setdiff1d(baddies, joining_exceptions)
        if baddies is None:
            continue
        
        bad_order = get_bad_order(transcript_cov, baddies, good_prefix="sscl")
        
        print("\n###################################")

    clear_output()
    
with open("basic_overlap_filter.json", "w") as handle:
    json.dump({"exceptions": list(joining_exceptions), "baddies": list(joining_transcripts)}, handle)

with open("basic_overlap_filter.json", "r") as handle:
    d = json.load(handle)
    joining_exceptions = set(d["exceptions"])
    joining_transcripts = set(d["baddies"])

Now I want to check that the excluded and included entries are actually ok.
Mistakes happen.

gtf[gtf["transcript"].isin(joining_transcripts | joining_exceptions)].head()

check_done = set()
check_responses = []

def get_check_response(mapper, subdf=None, linewidth=80):
    response = None
    # The unnecessary groupby is so that the order is the same as the plots.
    transcripts_tmp = set(subdf["transcript"])
    transcripts = list()

    for t in sorted(transcripts_tmp):
        transcripts.append([t, mapper[t], mapper[t]])

    while response not in ("y", "n"):
        print(f"\n")
        print("Please decide if your previous decisions were ok.")
        print("Enter [y]=yes/keep all, [n]=no/remove all anotations, or [e] to print exon structure.")
        print("Alternatively, you can enter a number to deal with an individual sequence.")

        max_mapper_len = max([len(j) for _, _, j in transcripts]) + 2
        max_id_len = max([len(i) for i, _, _ in transcripts])
        for h, (i, _, j) in enumerate(transcripts):
            print(f"- {j:<{max_mapper_len}}{i:>{max_id_len}}: {h}")

        response = input("\nresponse: ")
        response = response.strip().lower()

        try:
            response_int = int(response)
        except:
            response_int = None

        if response in ["n"]:
            return transcripts
        elif response in ["y"]:
            return transcripts
        elif response in ["e", "p", "print", "exons"]:
            assert subdf is not None, "If printing exons you need to have provided the subdf"
            print("\nExons:")
            plot_exons(subdf, step=20, mapper=mapper, linewidth=linewidth)

        elif (response_int is not None) and (response_int < len(transcripts)):
            response, prev, next_ = transcripts[response_int]

            print(f"\nEntry {response} was previously {prev}. What would you like to do with it?")

            response2 = "h"
            while response2 not in ("x", "d", "r"):
                response2 = input(f"\nEnter [x]=EXCEPTION, [d]=DELETED, [r]=RETAINED: ")
                response2 = response2.strip().lower()

            response2 = {"x": "EXCEPTION", "d": "DELETED", "r": "RETAINED"}[response2]
            transcripts[response_int][2] = response2
            mapper[transcripts[response_int][0]] = response2
    return




gids = gtf.loc[gtf["transcript"].isin(joining_transcripts | joining_exceptions), "gene_id"].unique()
gtf2 = gtf[gtf["gene_id"].isin(gids)]

for gene_id, subdf in gtf2.groupby("gene_id"):
    if gene_id in check_done:
        continue
    transcript_cov, gene_transcripts, baddies, easy_cases = is_bad(subdf, transcript_indices, coverages, good_prefix = "sscl")

    bad_order = get_bad_order(transcript_cov, baddies, good_prefix="sscl")
        
    current_baddie = bad_order[-1]

    print(f"# Processing: {gene_id}")

    #print(
    #    subdf["transcript type start end".split()]
    #        .sort_values("transcript start end".split())
    #        .set_index("transcript")
    #)
    
    mapper = {}
    for t in set(subdf["transcript"]):
        if t in joining_exceptions:
            mapper[t] = "EXCEPTION"
        elif t in joining_transcripts:
            mapper[t] = "DELETED"
        else:
            mapper[t] = "RETAINED"

    print("\nCoverage:")
    print(transcript_cov)
    print()
    plot_transcripts(subdf, step=100, mapper=mapper, linewidth=LINEWIDTH)

    r = get_check_response(mapper, subdf=subdf, linewidth=LINEWIDTH)
    check_responses.extend(r)
    check_done.add(gene_id)

    print("\n###################################")
    clear_output()

with open("basic_overlap_filter_check.json", "w") as handle:
    json.dump(check_responses, handle)

with open("basic_overlap_filter_check.json", "r") as handle:
    check_responses = json.load(handle)

responses = {
    tid
    for tid, _, decision
    in check_responses
    if decision == "DELETED"
}

Ok so i've curated a bunch of genes that are causing joining.
Now I need to assign new gene ids

## Another check for overlapping genes.

How many transcripts do we now have?

gtf.loc[~gtf["transcript"].isin(responses), "transcript"].nunique()

gtf2 = gtf.loc[~gtf["transcript"].isin(responses),].reset_index(drop=True)
gtf2.head()

From here it's relatively easy to find overlapping sets.

itrees = defaultdict(IntervalTree)
transcript_indices = dict()
indices_transcript = list()

for i, row in gtf2[gtf2["type"] == "transcript"].reset_index().iterrows():
    transcript = row["transcript"]
    transcript_indices[transcript] = i
    indices_transcript.append(transcript)
    itrees[(row["seqid"], row["strand"])].add(Interval(row["start"], row["end"], data=transcript))

def reducer(x, y):
    if isinstance(x, str):
        x = {x}
    elif x is None:
        x = set()
        
    x.add(y)
    return x
    
for itree in itrees.values():
    itree.merge_overlaps(data_reducer=reducer, data_initializer=set())

new_genes = {}
i = 1

for itree in itrees.values():
    for interval in sorted(itree, key=lambda x: x.begin):
        gene_id = f"AS{i:0>5}"
        i += 1
        
        for tid in interval.data:
            new_genes[tid] = gene_id       

gtf2["new_gene_id"] = gtf2["transcript"].apply(new_genes.get)
gtf2.head()

Now i can find genes with multiple ref_gene_ids.

new_loci = gtf2[gtf2["ref_gene_id"].notnull()].groupby("new_gene_id")["transcript"].nunique()
new_loci[new_loci > 1]

gid = "AS00292"

plot_transcripts(gtf2[gtf2["new_gene_id"] == gid])
x = gtf2.loc[(gtf2["new_gene_id"] == gid) & pd.isnull(gtf2["ref_gene_id"]), "transcript"].unique()

print(x)

responses.update(x)

gid = "AS00563"

plot_transcripts(gtf2[gtf2["new_gene_id"] == gid])
x = gtf2.loc[(gtf2["new_gene_id"] == gid) & pd.isnull(gtf2["ref_gene_id"]), "transcript"].unique()
x

responses.add("MSTRG.916.1")

gid = "AS01432"

plot_transcripts(gtf2[gtf2["new_gene_id"] == gid])
x = gtf2.loc[(gtf2["new_gene_id"] == gid) & pd.isnull(gtf2["ref_gene_id"]), "transcript"].unique()
x

m = gtf2.loc[gtf2["transcript"] == "sscle_01g010320", "end"].max()
gtf2.loc[gtf2["transcript"] == "MSTRG.1116.4", "end"] = gtf2.loc[gtf2["transcript"] == "MSTRG.1116.4", "end"].apply(lambda x: min([x, m]))

plot_transcripts(gtf2[gtf2["new_gene_id"] == gid])

gid = "AS01603"

plot_transcripts(gtf2[gtf2["new_gene_id"] == gid])
x = gtf2.loc[(gtf2["new_gene_id"] == gid) & pd.isnull(gtf2["ref_gene_id"]), "transcript"].unique()
x

responses.update(x)

gid = "AS02952"

plot_transcripts(gtf2[gtf2["new_gene_id"] == gid])
x = gtf2.loc[(gtf2["new_gene_id"] == gid) & pd.isnull(gtf2["ref_gene_id"]), "transcript"].unique()
x

m = gtf2.loc[gtf2["transcript"] == "sscle_03g022720", "end"].max()
gtf2.loc[gtf2["transcript"] == "MSTRG.2500.1", "end"] = gtf2.loc[gtf2["transcript"] == "MSTRG.2500.1", "end"].apply(lambda x: min([x, m]))

gtf2[(gtf2["new_gene_id"] == gid) & (gtf2["type"] == "transcript")]

gid = "AS04530"

plot_transcripts(gtf2[gtf2["new_gene_id"] == gid])
x = gtf2.loc[(gtf2["new_gene_id"] == gid) & pd.isnull(gtf2["ref_gene_id"]), "transcript"].unique()
x

responses.update(["MSTRG.4104.1", "MSTRG.4104.4"])

gid = "AS05956"

plot_transcripts(gtf2[gtf2["new_gene_id"] == gid])
x = gtf2.loc[(gtf2["new_gene_id"] == gid) & pd.isnull(gtf2["ref_gene_id"]), "transcript"].unique()
x

responses.update(x)

gid = "AS06269"

plot_transcripts(gtf2[gtf2["new_gene_id"] == gid])
x = gtf2.loc[(gtf2["new_gene_id"] == gid) & pd.isnull(gtf2["ref_gene_id"]), "transcript"].unique()
x

responses.update(['MSTRG.5144.1', 'MSTRG.5144.2'])

gid = "AS06672"

plot_transcripts(gtf2[gtf2["new_gene_id"] == gid])
x = gtf2.loc[(gtf2["new_gene_id"] == gid) & pd.isnull(gtf2["ref_gene_id"]), "transcript"].unique()
x

responses.update(['MSTRG.5814.1', 'MSTRG.5814.4'])

gid = "AS07931"

plot_transcripts(gtf2[gtf2["new_gene_id"] == gid])
x = gtf2.loc[(gtf2["new_gene_id"] == gid) & pd.isnull(gtf2["ref_gene_id"]), "transcript"].unique()
x

responses.update(['MSTRG.6394.2', 'MSTRG.6394.5'])

gid = "AS09271"

plot_transcripts(gtf2[gtf2["new_gene_id"] == gid])
x = gtf2.loc[(gtf2["new_gene_id"] == gid) & pd.isnull(gtf2["ref_gene_id"]), "transcript"].unique()
x

responses.update(['MSTRG.7965.4', 'MSTRG.7965.7', 'MSTRG.7965.8'])

gid = "AS09975"

plot_transcripts(gtf2[gtf2["new_gene_id"] == gid])
x = gtf2.loc[(gtf2["new_gene_id"] == gid) & pd.isnull(gtf2["ref_gene_id"]), "transcript"].unique()
x

responses.update(['MSTRG.8454.1', 'MSTRG.8454.2', 'MSTRG.8454.3', 'MSTRG.8454.5'])

gid = "AS10013"

plot_transcripts(gtf2[gtf2["new_gene_id"] == gid])
x = gtf2.loc[(gtf2["new_gene_id"] == gid) & pd.isnull(gtf2["ref_gene_id"]), "transcript"].unique()
x

responses.update(x)

gid = "AS10209"

plot_transcripts(gtf2[gtf2["new_gene_id"] == gid])
x = gtf2.loc[(gtf2["new_gene_id"] == gid) & pd.isnull(gtf2["ref_gene_id"]), "transcript"].unique()
x

responses.update(x)

gid = "AS12354"

plot_transcripts(gtf2[gtf2["new_gene_id"] == gid])
x = gtf2.loc[(gtf2["new_gene_id"] == gid) & pd.isnull(gtf2["ref_gene_id"]), "transcript"].unique()
x

responses.update(x)

gid = "AS12826"

plot_transcripts(gtf2[gtf2["new_gene_id"] == gid])
x = gtf2.loc[(gtf2["new_gene_id"] == gid) & pd.isnull(gtf2["ref_gene_id"]), "transcript"].unique()
x

responses.add('MSTRG.10767.3')

gid = "AS13467"

plot_transcripts(gtf2[gtf2["new_gene_id"] == gid])
x = gtf2.loc[(gtf2["new_gene_id"] == gid) & pd.isnull(gtf2["ref_gene_id"]), "transcript"].unique()
x

responses.add("MSTRG.11301.3")

gid = "AS04709"

plot_transcripts(gtf2[gtf2["new_gene_id"] == gid])
x = gtf2.loc[(gtf2["new_gene_id"] == gid) & pd.isnull(gtf2["ref_gene_id"]), "transcript"].unique()
x

responses.update(['MSTRG.4385.1', 'MSTRG.4385.2', 'MSTRG.4385.5'])

gid = "AS05330"

plot_transcripts(gtf2[gtf2["new_gene_id"] == gid])
x = gtf2.loc[(gtf2["new_gene_id"] == gid) & pd.isnull(gtf2["ref_gene_id"]), "transcript"].unique()
x

responses.add("MSTRG.4491.1")

Now i'm looking for components with low coverage.

gtf3 = gtf2.loc[~gtf2["transcript"].isin(responses),].reset_index(drop=True)
gtf3.head()

itrees = defaultdict(IntervalTree)
transcript_indices = dict()
indices_transcript = list()

for i, row in gtf3[gtf3["type"] == "transcript"].reset_index().iterrows():
    transcript = row["transcript"]
    transcript_indices[transcript] = i
    indices_transcript.append(transcript)
    itrees[(row["seqid"], row["strand"])].add(Interval(row["start"], row["end"], data=transcript))

def reducer(x, y):
    if isinstance(x, str):
        x = {x}
    elif x is None:
        x = set()
        
    x.add(y)
    return x
    
for itree in itrees.values():
    itree.merge_overlaps(data_reducer=reducer, data_initializer=set())
    
new_genes = {}
i = 1

for itree in itrees.values():
    for interval in sorted(itree, key=lambda x: x.begin):
        gene_id = f"AS{i:0>5}"
        i += 1
        
        for tid in interval.data:
            new_genes[tid] = gene_id

gtf3["new_gene_id2"] = gtf3["transcript"].apply(new_genes.get)
gtf3.head()

new_loci = gtf3[gtf3["ref_gene_id"].notnull()].groupby("new_gene_id2")["transcript"].nunique()
new_loci[new_loci > 1]

check2_done = set()
check2_responses = []

def intersect(a, b):
    astart, aend = a
    bstart, bend = b
    
    assert astart < aend, "a"
    assert bstart < bend, "b"
    ostart = max([astart, bstart])
    oend = min([aend, bend])
    
    if ostart > oend:
        return 0
    else:
        return oend - ostart

def union(a, b):
    astart, aend = a
    bstart, bend = b
    
    assert astart < aend, "a"
    assert bstart < bend, "b"
    #ostart = min([astart, bstart])
    #oend = max([aend, bend])
    al = aend - astart
    bl = bend - bstart
    return (al + bl) - intersect(a, b)

def gene_coverage(subdf):
    subdf = subdf[subdf["type"] == "transcript"].reset_index()
    mat = np.zeros((subdf.shape[0], subdf.shape[0]))
    
    for i, rowi in subdf.iterrows():
        for j, rowj in subdf.iterrows():
            inter = intersect((rowi["start"], rowi["end"]), (rowj["start"], rowj["end"]))
            uni = union((rowi["start"], rowi["end"]), (rowj["start"], rowj["end"]))
            min_ = min([abs(rowi["start"] - rowi["end"]), abs(rowj["start"] - rowj["end"])])
            mat[i, j] = inter / min_

    mat = pd.DataFrame(mat, columns=subdf["transcript"], index=subdf["transcript"])
    return mat

for gene_id, subdf in gtf3.groupby("new_gene_id2"):
    if gene_id in check_done:
        continue
    elif subdf["transcript"].nunique() == 1:
        continue

    cov = gene_coverage(subdf)
    
    if np.all(cov.values > 0.9):
        continue
    

    print(f"# Processing: {gene_id}")
    mapper = {}
    
    for t in subdf["transcript"]:
        mapper[t] = "RETAINED"

    print("\nCoverage:")
    print(cov)
    print()
    plot_transcripts(subdf, step=100, mapper=mapper, linewidth=LINEWIDTH)

    r = get_check_response(mapper, subdf=subdf, linewidth=LINEWIDTH)
    check2_responses.extend(r)
    check2_done.add(gene_id)

    print("\n###################################")
    clear_output()

with open("basic_overlap_filter_check2.json", "w") as handle:
    json.dump(check2_responses, handle)

with open("basic_overlap_filter_check2.json", "r") as handle:
    check2_responses = json.load(handle)

responses = {
    tid
    for tid, _, decision
    in check2_responses
    if decision == "DELETED"
}

gtf4 = gtf3.loc[~gtf3["transcript"].isin(responses),].reset_index(drop=True)
gtf4.head()

itrees = defaultdict(IntervalTree)
transcript_indices = dict()
indices_transcript = list()

for i, row in gtf4[gtf4["type"] == "transcript"].reset_index().iterrows():
    transcript = row["transcript"]
    transcript_indices[transcript] = i
    indices_transcript.append(transcript)
    itrees[(row["seqid"], row["strand"])].add(Interval(row["start"], row["end"], data=transcript))

def reducer(x, y):
    if isinstance(x, str):
        x = {x}
    elif x is None:
        x = set()
        
    x.add(y)
    return x
    
for itree in itrees.values():
    itree.merge_overlaps(data_reducer=reducer, data_initializer=set())
    
new_genes = {}
i = 1

for itree in itrees.values():
    for interval in sorted(itree, key=lambda x: x.begin):
        gene_id = f"AS{i:0>5}"
        i += 1
        
        for tid in interval.data:
            new_genes[tid] = gene_id

gtf4["new_gene_id3"] = gtf4["transcript"].apply(new_genes.get)
gtf4.head()

def jaccard(t1, t2):
    return (len(t1.intersection(t2)) / len(t1.union(t2)))

def jaccard_cov(df1, df2):
    one = set()
    two = set()
    
    for i, row in df1[df1["type"] == "exon"].iterrows():
        one.update(range(row["start"], row["end"]))
        
    for i, row in df2[df2["type"] == "exon"].iterrows():
        two.update(range(row["start"], row["end"]))
    
    return jaccard(one, two)


def gene_similarities(subdf):
    subdf_transcripts = subdf[subdf["type"] == "transcript"].drop_duplicates().reset_index()
    mat = pd.DataFrame(
        np.zeros((subdf_transcripts.shape[0], subdf_transcripts.shape[0])),
        columns=subdf_transcripts["transcript"],
        index=subdf_transcripts["transcript"]
    )    

    for i, subdfi in subdf.groupby("transcript"):
        for j, subdfj in subdf.groupby("transcript"):
            c = jaccard_cov(subdfi, subdfj)

            mat.loc[i, j] = c

    return mat

def get_check_response_jaccard(mapper, subdf=None, linewidth=80):
    response = None

    transcripts_li = sorted(subdf["transcript"].unique())
    transcripts = {t: "RETAINED" for t in transcripts_li}

    while response not in ("y", "n"):
        print(f"\n")
        print("Please decide if your previous decisions were ok.")
        print("Enter [y]=yes/keep all, [n]=no/remove all anotations, [f] to flag for followup, [d] to show the dataframe, or [e] to print exon structure.")
        print("Alternatively, you can enter a number to deal with an individual sequence.")

        max_mapper_len = max([len(v) for k, v in transcripts.items()]) + 2
        max_id_len = max([len(k) for k, v in transcripts.items()])
        for h, (k, v) in enumerate(transcripts.items()):
            print(f"- {v:<{max_mapper_len}}{k:>{max_id_len}}: {h}")

        response = input("\nresponse: ")
        response = response.strip().lower()

        try:
            response_int = int(response)
        except:
            response_int = None

        if response in ["n"]:
            return {k: "DELETED" for k, v in transcripts.items()}
        elif response in ["y"]:
            return transcripts
        elif response in ["f"]:
            return {k: "FOLLOWUP" for k, v in transcripts.items()}
        elif response in ["d"]:
            assert subdf is not None, "no dataframe"
            print("Dataframe")
            print(subdf)
        elif response in ["e", "p", "print", "exons"]:
            assert subdf is not None, "If printing exons you need to have provided the subdf"
            print("\nExons:")
            plot_exons(subdf, step=20, mapper=mapper, linewidth=linewidth)

        elif (response_int is not None) and (response_int < len(transcripts)):
            response = transcripts_li[response_int]
            prev = transcripts[response]

            print(f"\nEntry {response} was previously {prev}. What would you like to do with it?")

            response2 = "h"
            while response2 not in ("x", "d", "r"):
                response2 = input(f"\nEnter [x]=EXCEPTION, [d]=DELETED, [r]=RETAINED, [f]=FOLLOWUP: ")
                response2 = response2.strip().lower()

            response2 = {"x": "EXCEPTION", "d": "DELETED", "r": "RETAINED", "f": "FOLLOWUP"}[response2]
            transcripts[response] = response2
            mapper[response] = response2
    return

done = set()

mapper = {t: "RETAINED" for t in gtf4["transcript"].unique()}
jaccard_checks = dict()

non_overlapping = set()

for gene_id, subdf in gtf4.groupby("new_gene_id3"):
    if subdf["transcript"].nunique() == 1:
        continue
    sim = gene_similarities(subdf)

    if np.any(sim.values == 0):
        non_overlapping.add(gene_id)

    if gene_id in done:
        continue

    if np.any(sim.values < 0.4):
        print("\n", gene_id)
        print(sim)
        plot_exons(subdf, mapper=mapper, linewidth=160)
        print()

        responses = get_check_response_jaccard(mapper, subdf=subdf, linewidth=160)
        jaccard_checks.update(responses)
        clear_output()
    else:
        jaccard_checks.update({tid: mapper[tid] for tid in subdf["transcript"].unique()})

    done.add(gene_id)

with open("basic_overlap_filter_jaccard.json", "w") as handle:
    json.dump(jaccard_checks, handle)

with open("basic_overlap_filter_jaccard.json", "r") as handle:
    jaccard_checks = json.load(handle)

I happen to know that no transcripts additional genes needed to be excluded by the jaccard filter, but we did have some transcripts that existed entirely within the introns of other genes.
These are those genes:

non_overlapping

We'll mark them.

gtf_final = gtf4.copy()
gtf_final.drop(columns=["new_gene_id", "new_gene_id2", "ref_gene_id", "gene_id"], inplace=True)

gtf_final_order = (
    gtf_final
    [gtf_final["type"] == "transcript"]
    .groupby("new_gene_id3")
    [["seqid", "start", "end"]]
    .apply(lambda x: pd.Series({"seqid": x.iloc[0, 0], "start": x["start"].min(), "end": x["end"].max()}))
    .reset_index()
    .sort_values(["seqid", "start", "end"])
    .reset_index()
)

gtf_final_gene_mapper = {
    j: f"AS{i:0>5}" for i, j in enumerate(gtf_final_order["new_gene_id3"])
}

gtf_final["gene_id"] = gtf_final["new_gene_id3"].apply(gtf_final_gene_mapper.get)

gtf_final.sort_values(["seqid", "start", "end", "gene_id"], inplace=True)

gtf_final["orig_transcript_id"] = gtf_final["transcript"].apply(lambda x: x if x.startswith("sscle") else pd.NA)
gene_id_mapper = gtf_final[gtf_final["orig_transcript_id"].notnull()][["gene_id", "orig_transcript_id"]].drop_duplicates().set_index("gene_id")["orig_transcript_id"].to_dict()

gtf_final["orig_gene_id"] = gtf_final["gene_id"].apply(lambda x: f"gene-{gene_id_mapper.get(x, pd.NA)}")
gtf_final["has_intronic_nonoverlapping"] = gtf_final["new_gene_id3"].apply(lambda x: x in non_overlapping)

def mapper(row):
    out = []
    out.append(f'transcript_id "{row["transcript"]}"')
    out.append(f'gene_id "{row["gene_id"]}"')

    if pd.notnull(row["orig_gene_id"]):
        out.append(f'orig_gene_id "{row["orig_gene_id"]}"')
    
    if pd.notnull(row["orig_transcript_id"]):
        out.append(f'orig_transcript_id "{row["orig_transcript_id"]}"')
    
    if row["has_intronic_nonoverlapping"]:
        out.append(f'has_intronic_nonoverlapping "true"')

    return "; ".join(out)
        
gtf_final["attributes"] = gtf_final.apply(mapper, axis = 1)
gtf_final

gtf_final.to_csv("stringtie_gtf_processed.gtf", sep="\t", header=False, index=False, na_rep=".", )

(
    gtf_final
    ["seqid source type start end score strand phase attributes".split()]
    .to_csv("stringtie_gtf_processed.gtf", sep="\t", header=False, index=False, na_rep=".", quoting=3)
)
