==> 03-RNAseq/09-enrichment.Rmd <==
---
title: "09-enrichment.Rmd"
output: html_document
date: "2023-07-11"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r}
library(tidyverse)
library(UpSetR)
```

```{r}
func <- readr::read_tsv("input/protein_functions_long.tsv", na = ".")
head(func)
```

```{r}
eff <- readr::read_tsv("input/predicted_protein_functions.tsv", na = ".")
head(eff)
```

```{r}
published <- readr::read_tsv("input/published_sclero_genes.tsv")
head(published)
```



```{r}
ipvsiv <- list()
for (sheetname in readxl::excel_sheets("output/ipvsiv_meta.xlsx")) {
  ipvsiv[[sheetname]] <- readxl::read_xlsx("output/ipvsiv_meta.xlsx", sheet = sheetname)
}
```

```{r}
mapper <- function(df) {
  return(df %>% dplyr::select(id, host1, meta))
}

ipvsiv_dge <- do.call(rbind, lapply(ipvsiv, mapper)) %>%
  pivot_wider(id_cols = "id", names_from = "host1", values_from = "meta") %>%
  column_to_rownames("id")

ipvsiv_dge[is.na(ipvsiv_dge)] <- FALSE
ipvsiv_dge <- ipvsiv_dge[rowSums(ipvsiv_dge) > 0, ]

ipvsiv_dge <- as.matrix(ipvsiv_dge)

heatmap(1.0 * ipvsiv_dge[rowSums(ipvsiv_dge) > 1, ])
```

How many genes are DGE in all hosts.

```{r}
sum(apply(ipvsiv_dge, MARGIN = 1, FUN = all))
```

So what's common to most of the DGE sets?

```{r}
mapper <- function(df) {
  return(df %>% dplyr::filter(meta) %>% dplyr::select(id) %>% .[[1]])
}

ipvsiv_clist <- lapply(ipvsiv, mapper)

upset(fromList(ipvsiv_clist), nsets = length(ipvsiv_clist), order.by = "freq")
```
Here we're looking at genes DGE in at least 9 of 11 hosts.

```{r}
common_cands <- left_join(
  as.data.frame(ipvsiv_dge) %>% rownames_to_column("name") %>% mutate(n = rowSums(across(where(is.logical)))) %>% select(name, n) %>% filter(n > 9),
  eff %>% dplyr::select(name, effector_score, deeploc2_preds, effectordb_name, swissprot_protein_name, go_name, ipr_name),
  by = "name"
)

common_cands[order(common_cands$effector_score, decreasing = TRUE), ]
```

Ok. So as i expected the top candidates are mostly CAZymes.


```{r}
mapper <- function(df) {
  return(df %>% mutate(dge = ifelse(meta, lfc_median, NA)) %>% dplyr::select(id, host1, dge))
}

ipvsiv_lfc <- do.call(rbind, lapply(ipvsiv, mapper)) %>%
  pivot_wider(id_cols = "id", names_from = "host1", values_from = "dge") %>%
  column_to_rownames("id")

ipvsiv_lfc <- ipvsiv_lfc[!apply(is.na(ipvsiv_lfc), MARGIN = 1, all), ]
ipvsiv_lfc[is.na(ipvsiv_lfc)] <- 0
ipvsiv_lfc <- as.matrix(ipvsiv_lfc)
heatmap(ipvsiv_lfc)
```


```{r}
mapper <- function(df) {
  return(df %>% mutate(dge = ifelse(meta, lfc_median, NA)) %>% dplyr::select(id, host1, dge))
}

ipvsiv_lfc <- do.call(rbind, lapply(ipvsiv, mapper)) %>%
  pivot_wider(id_cols = "id", names_from = "host1", values_from = "dge") %>%
  column_to_rownames("id")

ipvsiv_lfc <- ipvsiv_lfc[!apply(is.na(ipvsiv_lfc), MARGIN = 1, all), ]

ipvsiv_eff <- left_join(ipvsiv_lfc %>% rownames_to_column("name"), eff, by = "name")
ipvsiv_eff <- ipvsiv_eff[order(ipvsiv_eff$effector_score), ]
readr::write_tsv(ipvsiv_eff, file = "~/test.tsv", na = "-")
```



```{r}
hostvshost <- list()
for (sheetname in readxl::excel_sheets("output/hostvshost_meta.xlsx")) {
  hostvshost[[sheetname]] <- readxl::read_xlsx("output/hostvshost_meta.xlsx", sheet = sheetname)
}
```





```{r}
mapper <- function(df) {
  return(df %>% dplyr::select(id, contrast_group, meta))
}

hostvshost_dge <- do.call(rbind, lapply(hostvshost, mapper)) %>%
  pivot_wider(id_cols = "id", names_from = "contrast_group", values_from = "meta") %>%
  column_to_rownames("id")

hostvshost_dge[is.na(hostvshost_dge)] <- FALSE
hostvshost_dge <- hostvshost_dge[rowSums(hostvshost_dge) > 0, ]

hostvshost_dge <- as.matrix(hostvshost_dge)

heatmap(1.0 * hostvshost_dge[rowSums(hostvshost_dge) > 1, ])
```

```{r}
mapper <- function(df) {
  return(df %>% dplyr::filter(meta) %>% dplyr::select(id) %>% .[[1]])
}

hostvshost_clist <- lapply(hostvshost, mapper)
```


```{r}
rcom_names <- names(hostvshost_clist) %>% base::Filter(function(x) {endsWith(x, "Rcom")}, .)
upset(fromList(hostvshost_clist[rcom_names]), nsets = length(rcom_names))
```

```{r}
hann_names <- names(hostvshost_clist) %>% base::Filter(function(x) {endsWith(x, "Hann")}, .)
upset(fromList(hostvshost_clist[hann_names]), nsets = length(hann_names))
```

```{r}
hann_names <- names(hostvshost_clist) %>% base::Filter(function(x) {endsWith(x, "Bvulvul")}, .)
upset(fromList(hostvshost_clist[hann_names]), nsets = length(hann_names))
```

```{r}
hann_names <- names(hostvshost_clist) %>% base::Filter(function(x) {endsWith(x, "Pvul")}, .)
upset(fromList(hostvshost_clist[hann_names]), nsets = length(hann_names))
```

```{r}
hann_names <- names(hostvshost_clist) %>% base::Filter(function(x) {endsWith(x, "Atha")}, .)
upset(fromList(hostvshost_clist[hann_names]), nsets = length(hann_names))
```

```{r}
hann_names <- names(hostvshost_clist) %>% base::Filter(function(x) {endsWith(x, "Bole")}, .)
upset(fromList(hostvshost_clist[hann_names]), nsets = length(hann_names))
```
